Added hook for deleting and delete unnecesary deletion modal from details
