+++ shared/hooks/index.js
@@ -9,3 +9,4 @@ export * from './BackendAPI/index';
 export * from './useCreateServiceBindingUsage';
 export * from './useProtectedResources';
 export * from './useCustomFormValidator';
+export * from './useDeleteResource';
